Fix(ts): Fix aws-cdk-lib versioning.
There is a issue with package.json version where some were to grab
latest, any version, pinned to certain that would cause sparatic issues
with building at certain times. Changed all of the packages that by
default cdk uses to follow the same syntax. Working on workflow that
will auto update to the same versions and syntax as cdk provides on
init.