Added sample stack for event sources mapped asynchronous lambda functions
It create a SNS topic, a SQS queue, and a EventBridge rule that trigger asynchronous lambda functions.
The source of event is the Amazon ECS events that are published into the default EventBridge event bus.
We use this event to push messages to the SNS topic and to the SQS queue.